requirements.txt
Django~=5.0.7
webcolors~=24.6.0
django-debug-toolbar~=4.4.6
django-ckeditor~=6.7.1